requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==0.61.1
django-allow-cidr==0.7.1
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
packaging==24.2
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rest-condition==1.0.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.8.2
